+++ app/src/main/java/com/example/woomansi/ui/main/Main3Fragment.java
@@ -1,27 +1,28 @@
-package com.example.woomansi;
+package com.example.woomansi.ui.main;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class Main1Fragment extends Fragment {
+import androidx.fragment.app.Fragment;
 
-    public Main1Fragment() {
+import com.example.woomansi.R;
+
+public class Main3Fragment extends Fragment {
+
+    public Main3Fragment() {
         // Required empty public constructor
     }
 
-    public static Main1Fragment newInstance() {
-        return new Main1Fragment();
+    public static Main3Fragment newInstance() {
+        return new Main3Fragment();
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_main1, container, false);
+        return inflater.inflate(R.layout.fragment_main3, container, false);
     }
 }
