+++ app/src/main/java/com/example/woomansi/ui/main/Main4Fragment.java
@@ -1,4 +1,4 @@
-package com.example.woomansi;
+package com.example.woomansi.ui.main;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.woomansi.R;
 
 public class Main4Fragment extends Fragment {
