+++ app/src/main/res/layout/fragment_main1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main1Fragment">
+    tools:context=".ui.main.Main1Fragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
