+++ app/src/main/res/layout/fragment_main2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main2Fragment">
+    tools:context=".ui.main.Main2Fragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
